Segundo cambio
Se han hecho la corrección en las líneas de código 15 y 16 tanto como en las 45 y 46, pues se ha realizado un cambio en el constructor tanto como en los métodos, para que estos llamen y reciban un argumento de tipo entero en ambos.

+++ src/main/java/classroom/Asignatura.java
@@ -12,8 +12,8 @@ public class Asignatura {
         this("Sin nombre",0,0,Tipo.FUNDAMENTACION);
     }
 
-    public Asignatura(int codigoInterno,Tipo tipo) {
-        this("Sin nombre", codigoInterno, 0,tipo);
+    public Asignatura(double codigoInterno,Tipo tipo) {
+        this("Sin nombre", (int) codigoInterno, 0,tipo);
     }
 
     public Asignatura(int codigoExterno,Tipo tipo) {
@@ -42,8 +42,8 @@ public class Asignatura {
         this.codigoInterno = codigoInterno;
     }
 
-    public void cambiarDatos(int codigoExterno) {
-        this.codigoExterno = codigoExterno;
+    public void cambiarDatos(double codigoExterno) {
+        this.codigoExterno = (int) codigoExterno;
     }
 
     public void cambiarDatos(String nombre) {
